Make minor change to prevent formatting carry-over
The addition of "\let\selectfont=\relax" on line 318 prevents formatting defined
within the "section heading" environment (for example, if the user selects a
different font family for their heading compared to the rest of their document)
from overwriting the "\cftsecfont" command. This allows the user to explicitly
define the font formatting of the table of contents independently of the
appearance of the section headings in the document body.

It should be noted that this solution resolves my particular case, but I'm sure
there is a more elegant and complete solution to the problem which I haven't
thought of.